requirements.txt
flask==3.1.0
line-bot-sdk==3.16.2
openai==1.66.3
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
pytz==2025.2